+++ example/reflection_factory_bridge_example.reflection.g.dart
@@ -1,6 +1,6 @@
 //
 // GENERATED CODE - DO NOT MODIFY BY HAND!
-// BUILDER: reflection_factory/1.0.14
+// BUILDER: reflection_factory/1.0.15
 // BUILD COMMAND: dart run build_runner build
 //